Send parent folder permissions along with workspace file metadata
The folder permissions are sent even if there is no workspace file for
the folder. This will make possible to disable the rich workspaces UI if
creating files is not enabled in a folder.

Rich workspaces are disabled for public share pages, so the folder
permissions are not sent in that case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>