detect local links and open in Viewer right away.
This used to be working but was broken
by having full hrefs rather than just the query part.

Now we check if the domain remains the same
and if we have a `dir` param in the query
and a `relPath` param in the fragment.

This should be a pretty good check for links
that we can open in the viewer right away.

Signed-off-by: Azul <azul@riseup.net>