fix: resolve relative image path based on currentDirectory option
This approach is more reliable because it does not rely on the files app.
Therefor it also works in `talk` and `collectives`.
Fixes #1780.

Use the currentDirectory to determine dav paths for gif files
and display the animated gif via dav rather than the preview
that lacks animation.

Also display webp files with dav urls. Fixes #1655.

Signed-off-by: Azul <azul@riseup.net>