Remove superfluous scrollbar caused by menububble
When the menububble is prepared
it is rendered at the bottom of the page and hidden.
The `margin-bottom: -3px` causes it to create a scrollbar.

Once the menububble is shown
it is positioned on the screen
and the scrollbar is gone.
So only apply the `margin-bottom: -3px`
when the menububble is shown.

Signed-off-by: Azul <azul@riseup.net>