Update mark input/paste rules to tiptap v2 regular expressions
The most important change is that a mark is added only if you are at
the beginning of a block or there is a space before it.

This especially fixes accidently formatting parts of a string as italic
when it contains two underscores (often true in URLs or emoji names).

This change can be reverted when we migrated to tiptap v2, but it's
a low hanging fix for an annoying bug, and the tiptap v2 migration might
take a bit longer.

See Philipp Kuehns comment at https://discuss.prosemirror.net/t/4230/3.

Signed-off-by: Jonas Meurer <jonas@freesources.org>