fix: use file id to identify readme for workspace
In public shares data did not include a name if i remember correctly.

When the Readme gets created it will trigger a PROPFIND
but that will only look for the Readme itself - not the entire directory.
Also detect that Readme so the view gets updated
and shows the rich workspace when a Readme.md is created.

This covers the main usecase that people click on the empty workspace.
It does not cover creating a localized readme file
or one named README.md or so.

We also did not cover that before - so this seems okay.

Signed-off-by: Max <max@nextcloud.com>