refactor: render Markdown in table nodes
Each node is responsible for rendering its markdown content.

This splits the responsibility and simplifies the `toMarkdown` functions a lot.
At the same time it makes it harder to beautify the entire table
because every cell only knows about itself - not about the rest of the column.

To work around this we can introduce padding attributes for TableCell and TableHeader.
These need to be updated when the content of the table changes.
They will also allow us to preserve the original padding from a markdown file.

Signed-off-by: Max <max@nextcloud.com>