fix: keep menubar in one line - prioritize entries
Menubar entries with submenus cannot be moved to the more-button.
Prioritize them when picking the icons for the toolbar.
If they still do not fit - do not add them to the more button either
they do not work there.

This makes sure that the menubar fits in one line and does not overflow.

Still prioritizing `undo` and `bold` over `emoji` and `callout`.
The former ones are need fairly frequently while the latter not so often.
In addition mobile phones have an emoji picker in their text entry.

Signed-off-by: Max <max@nextcloud.com>