Fix cypress scripts for local run + Add documentation
Fix the scripts for local cypress runs,
they need to be run within the `npm run` environment,
as the nextcloud webpack config requires the `npm_package_name`
variable to be set.

Also a up-to-date nextcloud instance is required for testing
text at the current git head, so updated the dockerfile
and docker-compose files to create a nextcloud image from current git head.

Additionally updated the tests to use cypress 10.x and
added missing documentation on how to run the cypress tests.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>