test: retry failed cypress tests twice
This will hopefully make the CI more stable
while allowing us to also see which tests are flaky
in the cypress ui.

Some tests may not be retryable yet.
So they will always fail on a second attempt because of remains of a first attempt.
We can tackle that on a test by test basis later.

Signed-off-by: Max <max@nextcloud.com>