BaseReader: Use a factory pattern for creating the extension list
Provide functions are neither called on the providing object nor
on the injecting object, so using `this` within provided functions
will refer to the global `this` for which `$emit` is not defined.
This uses a factory pattern so that `this` is bound to the injecting
object.

Also forwarding the `click-link` event is needed.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>