Do not remove escaping from brackets, fixes #2703
Removed the erroneous replacement of escaped brackets with
plain brackets as this breaks valid markdown.
Added a test case for this.

Switch to @hedgedoc/markdown-it-task-lists to fix
parsing of task lists, namely allowing any whitespace
inside the brackets, as by the github markdown specs.
Doing this as markdown-it-task-lists seems orphaned
while hedgedoc is maintained and the bug is fixed there.
This required adjusting the task-list testcase as the output is slightly different.

Removed two invalid testcases for task lists, as the specs only
allow exactly one `x`, `X` or whitespace character within the brackets.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>