Better styling of Front Matter + input rule
Added an input rule to add a front matter node if `---` is typed
in the very beginning of the document.
If it is typed somewhere else the default behavior is still
to insert a `hr` node.

Added a border on the left of the front matter, like the callouts
and blockquote, and added a title for the element.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>

a