fix: start from last saved version
We retrieve steps from version on only.
Since yjs will ask for missing steps anyway
there is no need to replay the entire history of steps.

Instead we rely on the local yjs doc to ask for changes
and others to send it a compressed update since version 0.

This prevents replaying SyncStep1 messages
which only caused new authors to send lots of replies.

Signed-off-by: Max <max@nextcloud.com>