Fix pressing the enter key removes heading content
`div` is not an allowed child element of a `hX` tag,
this leads to invalid behavior on firefox so we need to use the `span` tag
as the NodeView content container.

Also using `tabindex=-1` on the NodeView will trigger weird behavior when
clicking inside the heading and pressing the enter key. It will split the node
but also copy the heading anchor link of the NodeView.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>