Remove trailing newline from code blocks
Markdown-it serialized code blocks to `<pre><code>...\n</pre></code>`,
which leads to trailing newlines in the code blocks as rendered by
 Tiptap.

Let the Tiptap CodeBlock node remove the trailing newline in parseHTML
in order to avoid this.

There's another issue with trailing newlines that got added to code
blocks on purpose being removed by prosemirror-markdown, but that's not
affected by this change.

Fixes: #2344

Signed-off-by: Jonas <jonas@freesources.org>