Don't empty list of headings in onDestroy
When switching pages in Collectives, `onDestroy` from the old editor
instance happens *after* `onCreate` of the new editor instance. Since
both share the same vuex store, this leads to a race condition where the
old instance clears the list of headings of the new instance.

Since we always set headings during `onCreate` anyway, there's no need
to clear the headings in `onDestroy`, so let's simply get rid of it.

Signed-off-by: Jonas <jonas@freesources.org>