fix: flaky cypress test in workspace test
Fix the selector for main menu entries
to exclude entries of dangling submenus.

Whenever the strikethrough action was inside the remaining actions
the `formats text` test would fail.

`getMenuEntry` always returned main menu entries directly.
For entries of submenus it should open that submenu
and then return the entry in question.

When attempting to click the strikethrough action
it would detect the action-entry in the remain actions drop down menu
as if it was inside the main menu.
It would then return it without opening the remain menu.

The dangling menu with the action was hidden.
But since `click` was called with `force: true`
it would still happen.
However it had no impact.
The click just happened in the text field itself
where the detached menu was located but not clickable.

Fixing the selector to distinguish
between main menu actions and remaining actions
also allows us to drop some other `force` options.

Signed-off-by: Max <max@nextcloud.com>