Catch shiftKey when pasting with middle click
We distinguish between <Ctrl>-<V> and <Ctrl>-<Shift>-<V> when pasting
text to decide whether the input should be treated as markdown or
plaintext.

On Linux systems, it's also possible to paste text with a middle mouse
click. Let's consider there as well whether the <Shift> key is down.

Sidenote: with shift key down, the handleClick and handleClickOn events
are not called, so we have to go through the DOM events directly.

Signed-off-by: Jonas <jonas@freesources.org>