fix(SessionApi): only send content if no session is open
If a session is open - even if no state has been saved yet
the content of the document will be recovered from the steps.

In this case do not send the content in the response to the create request
as it would be duplicated:
* once created when the editor is initialized with the content
* then again when the steps are replayed.

Include tests for this corner case and another one.

Signed-off-by: Max <max@nextcloud.com>