fix(Session): race condition in session creation
When one user creates a session but never pushes the initial steps
send the content of the file to the next user as well.

This still opens the door for another race condition:
If the first user continues to establish the session afterwards
the initial content will be loaded twice.

Will tackle this in another commit

Signed-off-by: Max <max@nextcloud.com>