fix(autosave): only autosave after actual changes happened
With the yjs backend loading the file will already trigger a push.
In the past we also had the problem that some extensions
triggered pushes even without changes to the content.

Use the state of the document history to determine if autosave is needed.
Do not save if there is no change to undo / redo.

Also move the autosave functionality into the SyncService.
This reduces the complexity of the PollingBackend further
and makes it readily accessible from the Editor.

Signed-off-by: Max <max@nextcloud.com>