fix(DirectEditing): delay the initialSession so hooks can register
In Direct Editing we seed the initial session.
Therfore we can avoid some requests.
However if we use the data right away the onOpened function is called
before the hooks of the yjs websocket have even been registered.

Use a Promise.resolve instead to delay the opening
so hooks are registered first.

Signed-off-by: Max <max@nextcloud.com>