fix: load fresh session if none are remaining
When creating a new session
check for all existing sessions
to take those into account
that are only preserved because of unsaved changes.

It there still aren't any sessions
make sure to load a fresh editing session
even if there are unsaved steps.
These steps most likely originate from a race condition
when closing the last editing session and pushing steps at the same time.

We were seeing empty editing sessions in production
while the markdown files had some content.

This was happening because there were some steps still around
even though all sessions had been closed.

When the last session was closed
steps were pushed at the same time.
The push passed the session check before the session was cleared.
The session data and the steps were cleared
and only then the new step was added.

This left the database in an inconsistent state:
* just a few late steps instead of a full history
* no sessions these steps would belong to

Instead of trying to prevent this race condition
detect this state and recover from it.

Sessions are sometimes aborted without close request in the middle of editing.
In this case all steps are present in the database
and the session also remains.

The session will not be cleaned up during `resetDocument`.
Instead `resetDocument` throws `DocumentHasUnsavedChagesException`.

We can detect this scenario because even `removeInactiveSessions`
also takes remaining steps into account and leaves sessions with steps alone.
Therefore sessions will remain in `apiService->create(...)`
and `freshSession` won't be set.

Steps are only cleaned up when the last active session is closed.
Even if a few steps manage to sneak in after the clean up
the session will be removed
and so `remainingSessions` will be empty during `create`.
Therefore `freshSession` will be set and the content will be send out.
The leftovers won't be cleaned up until the file has been saved again.
But that should be fine. They will not impact the editing.

Signed-off-by: Max <max@nextcloud.com>