fix(Viewer): remove outdated fix from stable16.
This broke the viewer resize and is not needed anymore.

In order to confirm this is not needed
i replaced the `beforeMount` handler with a `console.info` call
opened a text in the viewer and resized the window a few times.
The provided info only showed up once.
So the `ViewerComponent` was only mounted once as well.

Fixes #4099.

Signed-off-by: Max <max@nextcloud.com>