fix(sync): Save even if versions match
Fixes #3404.

Apply saves even if the client and server version match.

The client version only reflects the steps which the client
has received back from the server.
It may leave out the steps the client just send itself.

So if the versions match - save the file to be sure to include changes from the client.

Signed-off-by: Max <max@nextcloud.com>