fix(links): Insert `/f/<fileId>` format for links to files
Get rid of the old `/path/?fileId=<id>` format. It never really worked
anyway because the paths were relative to the users' home directory.

Instead, insert links to files/folders as full URLs with the
`/f/<fileId>` format. This makes link previews work both in Text and
Collectives.

Rewrite links in the old format to the new format. But don't touch
absolute link (without origin) to the collectives app to not break
collectives-specific link handling to other pages.

Signed-off-by: Jonas <jonas@freesources.org>