fix(link): Insert full URLs for links to Nextcloud files
Full URLs have the advantage that they also work when clicking them
outside the Nextcloud context, e.g. in a local markdown editor.

Their main disadvantage is that they cause trouble on Nextcloud
instances with more than one hostname or after changing the hostname of
a Nextcloud instance.

The first use case (opening a file outside Nextcloud context) is more
likely and less of a corner case than the second (changing hostname), so
we decided to go for full URLs.

Signed-off-by: Jonas <jonas@freesources.org>