fix(LinkBubble): Use debounce instead of custom debounce implementation
Run `updateFromSelection()` immediately on first call and only delay
further runs.

This fixes a race condition with `this.#hadUpdateFromClick` which gets
set when updating from click and gets cleaned up after 200ms.

Fixes an issue with the bubble disappearing immediately when clicking on
a link after unfocussed browser and editor.

Signed-off-by: Jonas <jonas@freesources.org>