fix(sync): reuse open connection
Do not attempt to create a new connection
if there already is one and it is not closed.

If no messages are received for 30 seconds
yjs will open a new websocket.

Since we do not close the connection anymore from the websocket polyfill
we also do not need to open it.

If the network connection has gone down
creating a new connection will fail anyway.

Once it comes back we will know if the session is still valid.
Then we can either continue using it or reconnect.

This is part of #6050.

Signed-off-by: Max <max@nextcloud.com>