fix(ui): transparent background behind document status
The document status div is full width.
Its background hides the text behind it and scrollbars on the side.

Introduce a new .status-wrapper div that sets the background
and only has the size of the actual message.

Also move the message to the very bottom of the screen on mobile.

Signed-off-by: Max <max@nextcloud.com>