fix(LinkBubble): Read `editor.contentComponent` when it's ready
When initiating the editor extensions (i.e. in `addProseMirrorPlugins()`),
`editor.contentComponent` is not yet available. So let's read it when we
actually use it, when creating the tooltip in `LinkBubblePluginView`.

Before, a new Vue instance was initiated for each `LinkBubbleView.vue`
as the `parent` given to `VueRenderer()` was `undefined`. With this fix,
`LinkBubbleView.vue` becomes a child component of the root editor Vue
instance.

Signed-off-by: Jonas <jonas@freesources.org>