fix(Keymap): Add Backspace shortcut to undo input rules
The upstream Tiptap Keymap extension that we don't use (anymore) also
provides this.

It allows to undo the automatic changes from input rules directly after
they got applied by hitting <Backspace>.

Unfortunately this doesn't work for nodes that get changed by the
TextDirection extension due to a bug, but there's still plenty of input
rules where it works, so let's bring it in.

Signed-off-by: Jonas <jonas@freesources.org>