chore(cleanup): do not store nick after loading it
We load the guestName from localStorage,
send it to the server,
and then store it again.

That would make sense if the server changed it.
But i do not see why it would do that
if it did not do it the last time the name was changed and stored in localStorage.

Signed-off-by: Max <max@nextcloud.com>