fix(links): do not steal focus on initial loading
The `LinkBubbleView` is created and then mounted by `createTooltip`
in the `update` function of the `LinkBubblePluginView`.

This already happens during the initial rendering
as that is the first time the editor content is updated.

At this time the `href` is null and therefore `startEdit` is called
which in turn sets the focus in the entry field.

Instead set `edit` to `true` right away.

The behavior of `mod-k` is still inconsistent.
In particular it will not focus the link entry
if the `href` already was falsy - for example right after the start.

Signed-off-by: Max <max@nextcloud.com>