fix(createEditor): always provide editor.
Set content after providing the editor.

If the content contains links the editor renders `PreviewOptions`.

Rendering another component breaks the context of the setup function.
Namely `setCurrentInstance` gets called in the vue runtime
overwriting the current instance.

This leads to `provide` not working anymore as it has no current instance
to work with.

Signed-off-by: Max <max@nextcloud.com>