fix(sync): lastSavedVersion is version send with the request
The client sends a serialized version of the document
plus the document state
according to the steps it received until that point
(maybe plus local edits that were not synced yet).

lastSavedVersion should be the version the client had seen
at the time it initiated the save request
not the last version of any step the server saw
when receiving the request.

Otherwise when loading the document
or attempting to recover from missing steps
the steps in between will be left out.

Signed-off-by: Max <max@nextcloud.com>