fix(editor): restore default editor width while not breaking drag handle
This way the drag handle overflows the editor width just like all the
other action buttons left and right of the editor already did before.

To make sure the drag handle stays when leaving the editor container with
the pointer while trying to reach the drag handle, the latter gets an
extra `padding-right` to close the gap.

Since the link preview options three-dot menu is in that gap, it needs
a higher z-index to stay accessible.

FIxes: #7649
Fixes: nextcloud/collectives#2020

Signed-off-by: Jonas <jonas@freesources.org>