fix(sync): Reset document sessions on upgrade
A bug in synchronizing the document state might have resulted in a Yjs
document state with missing steps being persisted into the Ydoc that we
store on the server. If no client that had the missing step did an
autosave afterwards, future sync session members will never recover from
this situation.

We implemented several improvements recently that should result in less
out-of-sync situations where a step becomes missing. To start with fresh
editing sessions and not carry on with possibly unrecoverable document
states, let's reset all editing sessions with the next app update.

See #7692 for further context.

Signed-off-by: Jonas <jonas@freesources.org>