fix(sync): actually disable browser broadcast
It was still enabled as we mistyped the `disableBc` option.

We cannot use it right now as the y-websocket room name
does not include the base version etag.

Therefore different tabs may be in different sessions
and would risk syncing never the less.

Observed this when reloading the editor on an outdated session:
* Cleaning up the old yjs document took a while in the beforeDestroy callback.
* The new document was already created.
* The new y-websocket provider issued a sync step 1 via browser broadcast.
* The old y-websocket provider responded and send its full state.

Signed-off-by: Max <max@nextcloud.com>