fix(table): prevent nested tables
Check pasted content for possibly nested table when pasting to a table
cell. Show an alert if nested table is detected.

Add some extra logic to detect if just one table cell got selected and
copied, which results in a table with one row and cell in the clipboard.
Unwrap the cell content in this case and paste it.

Add `filterTransaction` to prevent applying a transaction that would
result in nested tables. This is meant as an extra low-level protection
against nested tables.

Signed-off-by: Jonas <jonas@freesources.org>